Fix init decorator
init is intended to be a decorator which may be called with either
zero or one arguments. In the first case, the argument passed to
init is the function to be decorated, so cmd_args ought to be None,
and in the second, the argument to init is the parser, so cmd_args
ought to be arg.parse_args(). This fixes the arity issue that
formerly arose when trying to run mesos-help (which used cli.init
without arguments).